Rename state and validator to channelstate and channelvalidator
