+++ packages/arb-validator/channelstate/unanimous.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package state
+package channelstate
 
 import (
 	"context"
